+++ packages/emitter/testcases/common/expected/lang/stackalloc/StackAlloc.cs
@@ -5,8 +5,8 @@ namespace TestCases.common.lang.@stackalloc
     {
         public static void run()
         {
-            global::System.Span<int> buffer = stackalloc int[(int)256];
-            buffer[0] = (int)42;
+            global::System.Span<int> buffer = stackalloc int[256];
+            buffer[0] = 42;
             global::System.Console.WriteLine(buffer[0]);
             global::System.Console.WriteLine(buffer.Length);
         }
